feat : [006-CUSTOMER-BALANCE] Customer 에 대한 잔액 조회 구현

+++ user-api/src/main/java/com/zerobase/cms/user/domain/model/Customer.java
@@ -35,7 +35,8 @@ public class Customer extends BaseEntity{
     private String verificationCode;
     private boolean verify;
 
-    private Long balance;
+
+    private Integer balance;
     public static Customer from(SignUpForm form) {
         return Customer.builder()
                 .email(form.getEmail().toLowerCase(Locale.ROOT))
